requirements/base.txt
Flask==1.1.1
Flask-RESTful==0.3.8
psycopg2-binary==2.8.5
python-dotenv==0.14.0
Flask-SQLAlchemy==2.4.1
Flask-Cors==3.0.8
Flask-User==1.0.2.2
email-validator==1.1.1
SQLAlchemy-Utils==0.36.8
marshmallow==3.7.1
bcrypt==3.2.0
